removed duplicate tests
